refactor: formatting and moved components inside indexPage
